% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aclust.R
\name{aclust}
\alias{aclust}
\title{Generate (perhaps approximate) dendrogram for heirarchical clustering of the rows of mat.}
\usage{
aclust(
  mat,
  method = "ward",
  distance = "euclidean",
  p = 2,
  pcLimit = 512,
  pcSize = 128,
  pcMaxRows = 2048,
  hclustMax = 25000
)
}
\arguments{
\item{mat}{A numeric matrix with row labels}

\item{method}{Agglomeration method passed to hclust}

\item{distance}{Distance metric passed to hclust}

\item{p}{Minkowski distance parameter passed to hclust}

\item{pcLimit}{Maximum number of columns before using principal components}

\item{pcSize}{Number of principal components to use}

\item{pcMaxRows}{Maximum number of rows to use for principle components}

\item{hclustMax}{Maximum number to rows to hclust at a time}
}
\value{
A dendrogram for the rows of mat
}
\description{
mat must be a numeric matrix with row labels.
mat will be column normalized before clustering. So no column can have constant values.
}
\details{
Approximate methods will be employed if:
\itemize{
\item the number of columns exceeds pcLimit (default 512):
Clustering will be performed using the first pcSize (default 128) principal components.
If the number of rows exceeds pcMaxRows (default 2048), the principal components will be
approximated using pcMaxRows randomly selected rows.
\item the number of rows exceeds hclustMax (default 25000), the rows will partitioned using
kmeans 2 as many times as required until each partition has at most hclustMax rows.
}

Rclusterpp::Rclusterpp.hclust will be used for (subsets of) the matrix with at most hclustMax rows.
The parameters method, distance, and p are passed to it without modification.
}
