% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale.R
\name{rescale}
\alias{rescale}
\title{Rescale a dendrogram based on a data matrix.}
\usage{
rescale(ddg, mat)
}
\arguments{
\item{ddg}{A dendrogram}

\item{mat}{A numeric matrix with row labels that include all the labels in ddg.}
}
\value{
A dendrogram rescaled so that node heights equal the total within-cluster differences in mat.
}
\description{
Recalculate the height of the dendrogram branches based on the within-
cluster sum of differences of the data within the numeric matrix mat.
All labels of ddg must correspond to rows in mat.
}
